Made catalog, books details, authorization, authantication, personal info

+++ PoolOfBooks/Models/Books.cs
@@ -10,7 +10,7 @@
         public int? id_category { get; set; }
         public string cover { get; set; } = "твердая";
         public int? in_stock { get; set; }
-        public float? price { get; set; }
+        public decimal? price { get; set; }
         public string status { get; set; } = "аренда";
         public int? count_was_read { get; set; }
